fix: typechain directory rename
