+++ .gitignore
@@ -3,7 +3,7 @@ artifacts
 cache
 
 # TypeChain
-typechain
+typechain-types
 
 # Logs
 logs
