+++ package.json
@@ -27,14 +27,14 @@
     "test": "mocha --timeout 10000 --exit --recursive --require ts-node/register test/**/*.test.ts"
   },
   "devDependencies": {
-    "@nomiclabs/hardhat-ethers": "^2.0.3",
-    "@nomiclabs/hardhat-etherscan": "^2.1.8",
-    "@nomiclabs/hardhat-waffle": "^2.0.1",
+    "@nomiclabs/hardhat-ethers": "^2.0.4",
+    "@nomiclabs/hardhat-etherscan": "^3.0.0",
+    "@nomiclabs/hardhat-waffle": "^2.0.2",
     "@openzeppelin/contracts": "^4.4.2",
     "@openzeppelin/contracts-upgradeable": "^4.4.2",
-    "@openzeppelin/hardhat-upgrades": "^1.12.0",
-    "@typechain/ethers-v5": "^7.0.1",
-    "@typechain/hardhat": "^2.3.0",
+    "@openzeppelin/hardhat-upgrades": "^1.14.0",
+    "@typechain/ethers-v5": "^9.0.0",
+    "@typechain/hardhat": "^4.0.0",
     "@types/bunyan": "^1.8.8",
     "@types/chai": "^4.2.21",
     "@types/chai-as-promised": "^7.1.4",
